% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{garamszegi_phy}
\alias{garamszegi_phy}
\alias{garamszegi_pois_missing}
\alias{garamszegi_repeat}
\alias{garamszegi_simple}
\alias{garamszegi_pois}
\alias{garamszegi_effect}
\title{data}
\format{
An object of class \code{phylo} of length 4.

An object of class \code{data.frame} with 200 rows and 3 columns.

An object of class \code{data.frame} with 1000 rows and 4 columns.

An object of class \code{data.frame} with 200 rows and 3 columns.

An object of class \code{data.frame} with 200 rows and 3 columns.

An object of class \code{data.frame} with 200 rows and 3 columns.
}
\usage{
garamszegi_phy

garamszegi_pois_missing

garamszegi_repeat

garamszegi_simple

garamszegi_pois

garamszegi_effect
}
\description{
#' Simple data from Garamszegi chapter 11
}
\keyword{datasets}
