% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme4_phylo_setup.R
\name{glFormula}
\alias{glFormula}
\title{hacked version of glFormula}
\usage{
glFormula(
  formula,
  data = NULL,
  family = gaussian,
  subset,
  weights,
  na.action,
  offset,
  contrasts = NULL,
  start,
  mustart,
  etastart,
  control = glmerControl(),
  phylonm,
  phyloZ,
  ...
)
}
\description{
hacked version of glFormula
}
