% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmTMB_phylo_setup.R
\name{glmmTMBphylo}
\alias{glmmTMBphylo}
\title{hacked version of glmmTMB}
\usage{
glmmTMBphylo(
  formula,
  data = NULL,
  family = gaussian(),
  ziformula = ~0,
  dispformula = ~1,
  weights = NULL,
  offset = NULL,
  contrasts = NULL,
  phyloZ = NULL,
  phylonm = NULL,
  na.action = na.fail,
  se = TRUE,
  verbose = FALSE,
  doFit = TRUE,
  control = glmmTMBControl(),
  REML = FALSE,
  map = NULL,
  sparseX = NULL
)
}
\description{
hacked version of glmmTMB
}
