% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme4_phylo_setup.R
\name{lFormula}
\alias{lFormula}
\title{hacked version of lme4::lFormula}
\usage{
lFormula(
  formula,
  data = NULL,
  REML = TRUE,
  subset,
  weights,
  na.action,
  offset,
  contrasts = NULL,
  control = lme4::lmerControl(),
  phylonm,
  phyloZ,
  ...
)
}
\arguments{
\item{formula}{mixed-model formula}

\item{data}{data frame}

\item{REML}{use restricted max likelihood?}

\item{subset}{subset expression}

\item{weights}{weights}

\item{na.action}{na.action}

\item{offset}{offset}

\item{contrasts}{contrasts}

\item{control}{control}

\item{phylonm}{name of phylogenetic grouping variable}

\item{phyloZ}{phylogenetic Z-matrix}
}
\description{
hacked version of lme4::lFormula
}
