% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme4_phylo_setup.R
\name{mkReTrms}
\alias{mkReTrms}
\alias{mkBlist}
\alias{glFormula}
\title{hacked version of lme4::mkReTrms}
\usage{
mkReTrms(bars, fr, phylonm, phyloZ, drop.unused.levels = TRUE)

mkBlist(x, frloc, phylonm, phyloZ, drop.unused.levels = TRUE)

glFormula(
  formula,
  data = NULL,
  family = gaussian,
  subset,
  weights,
  na.action,
  offset,
  contrasts = NULL,
  start,
  mustart,
  etastart,
  control = glmerControl(),
  phylonm,
  phyloZ,
  ...
)
}
\arguments{
\item{phylonm}{name of phylogenetic grouping variable}

\item{phyloZ}{phylogenetic Z-matrix}

\item{formula}{mixed-model formula}

\item{data}{data frame}

\item{subset}{subset expression}

\item{weights}{weights}

\item{na.action}{na.action}

\item{offset}{offset}

\item{contrasts}{contrasts}

\item{control}{control}
}
\description{
hacked version of lme4::mkReTrms

hacked version of mkBlist

hacked version of glFormula
}
