% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmTMB_phylo_setup.R
\name{mkTMBStrucphylo}
\alias{mkTMBStrucphylo}
\title{make TMB structure}
\usage{
mkTMBStrucphylo(
  formula,
  ziformula,
  dispformula,
  combForm,
  mf,
  fr,
  yobs,
  respCol,
  weights,
  contrasts = contrasts,
  size = NULL,
  family,
  se = NULL,
  phyloZ = phyloZ,
  phylonm = phylonm,
  call = NULL,
  verbose = NULL,
  ziPredictCode = "corrected",
  doPredict = 0,
  whichPredict = integer(0),
  REML = FALSE,
  map = NULL,
  sparseX = NULL
)
}
\description{
make TMB structure
}
