% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme4_phylo_setup.R
\name{phylo_glmm}
\alias{phylo_glmm}
\title{phylogenetic GLMM}
\usage{
phylo_glmm(
  formula,
  data,
  phylo,
  phylonm = NULL,
  phyloZ = NULL,
  control,
  family
)
}
\description{
phylogenetic GLMM
}
