% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme4_phylo_setup.R
\name{phylo_machinery}
\alias{phylo_machinery}
\alias{phylo.to.Z}
\title{construct a Z matrix from a phylogeny}
\usage{
phylo.to.Z(r, stand = FALSE)
}
\arguments{
\item{r}{a \code{phylo} object}

\item{stand}{standardize edge lengths by determinant of phylogenetic covariance matrix}
}
\description{
construct a Z matrix from a phylogeny
}
