% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme4_phylo_setup.R
\docType{data}
\name{mkReTrms}
\alias{mkReTrms}
\alias{mkBlist}
\alias{gl_tmp}
\title{hacked version of lme4::mkReTrms}
\format{
An object of class \code{character} of length 71.
}
\usage{
mkReTrms(bars, fr, phylonm, phyloZ, drop.unused.levels = TRUE)

mkBlist(x, frloc, phylonm, phyloZ, drop.unused.levels = TRUE)

gl_tmp
}
\arguments{
\item{phylonm}{name of phylogenetic grouping variable}

\item{phyloZ}{phylogenetic Z-matrix}
}
\description{
hacked version of lme4::mkReTrms

hacked version of mkBlist

hacked version of glFormula
}
\keyword{datasets}
