% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme4_phylo_setup.R
\name{phylo_lmm}
\alias{phylo_lmm}
\alias{phylo_glmm}
\title{phylogenetic linear mixed model}
\usage{
phylo_lmm(
  formula,
  data,
  phylo = NULL,
  phylonm = NULL,
  phylosp = phylonm[1],
  phyloZ = NULL,
  control,
  REML = FALSE
)

phylo_glmm(
  formula,
  data,
  phylo,
  phylonm = NULL,
  phylosp = phylonm[1],
  phyloZ = NULL,
  control,
  family
)
}
\arguments{
\item{formula}{mixed-model formula}

\item{data}{data frame}

\item{phylo}{phylogenetic tree in \code{\link{phylo}} format}

\item{phylonm}{name of phylogenetic grouping variables; if \code{phylosp} is not specified, the first element must match the name of the species variable in \code{data}}

\item{phylosp}{name of species variable in \code{data} (levels should match the tip/species names from the phylogeny)}

\item{phyloZ}{phylogenetic Z-matrix (see \code{\link{phylo.to.Z}}): optional, will be computed
internally from \code{phylo} if not specified here. (For large phylogenies that are going to
be used in multiple models it may make sense to compute the Z matrix first and pass it to
the modeling functions to avoid recomputing it each time.)
to compute \code{phyloZ} fi}

\item{control}{control}

\item{REML}{use restricted max likelihood?}

\item{family}{GLMM family}
}
\description{
phylogenetic linear mixed model

phylogenetic GLMM
}
