% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme4_phylo_setup.R
\name{phylo_lmm}
\alias{phylo_lmm}
\title{phylogenetic linear mixed model}
\usage{
phylo_lmm(
  formula,
  data,
  phylo,
  phylonm = NULL,
  phyloZ = NULL,
  control,
  REML,
  doFit
)
}
\description{
phylogenetic linear mixed model
}
